logic udpated
