+++ rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/activity/CovidJsonFormActivity.java
@@ -7,10 +7,12 @@ import android.content.pm.PackageManager;
 import android.graphics.Rect;
 import android.location.Location;
 import android.os.Bundle;
-import android.os.Looper;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
+
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
 
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationCallback;
@@ -24,13 +26,12 @@ import com.vijay.jsonwizard.fragments.JsonFormFragment;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import androidx.annotation.NonNull;
-import androidx.core.app.ActivityCompat;
 import io.ona.rdt.R;
 import io.ona.rdt.application.RDTApplication;
 import io.ona.rdt.fragment.CovidJsonFormFragment;
 import io.ona.rdt.util.CovidRDTJsonFormUtils;
 import io.ona.rdt.util.RDTJsonFormUtils;
+import io.ona.rdt.util.Utils;
 import timber.log.Timber;
 
 import static io.ona.rdt.util.CovidConstants.FormFields.LAST_KNOWN_LOCATION;
@@ -67,7 +68,7 @@ public class CovidJsonFormActivity extends RDTJsonFormActivity {
     @Override
     protected void onResume() {
         super.onResume();
-        RDTApplication.getInstance().verifyUserAuthorization();
+        Utils.verifyUserAuthorization();
     }
 
     @Override
