+++ rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/activity/CovidPatientProfileActivity.java
@@ -5,16 +5,16 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 
+import androidx.fragment.app.Fragment;
+import androidx.viewpager2.widget.ViewPager2;
+
 import com.google.android.material.tabs.TabLayout;
 import com.google.android.material.tabs.TabLayoutMediator;
 
 import org.json.JSONException;
 
-import androidx.fragment.app.Fragment;
-import androidx.viewpager2.widget.ViewPager2;
 import io.ona.rdt.R;
 import io.ona.rdt.adapter.ProfileFragmentAdapter;
-import io.ona.rdt.application.RDTApplication;
 import io.ona.rdt.contract.CovidPatientProfileActivityContract;
 import io.ona.rdt.domain.Patient;
 import io.ona.rdt.fragment.CovidPatientProfileFragment;
@@ -24,6 +24,7 @@ import io.ona.rdt.presenter.PatientProfileActivityPresenter;
 import io.ona.rdt.util.Constants;
 import io.ona.rdt.util.FormKeyTextExtractionUtil;
 import io.ona.rdt.util.RDTJsonFormUtils;
+import io.ona.rdt.util.Utils;
 import timber.log.Timber;
 
 /**
@@ -44,7 +45,7 @@ public class CovidPatientProfileActivity extends PatientProfileActivity implemen
     @Override
     protected void onResume() {
         super.onResume();
-        RDTApplication.getInstance().verifyUserAuthorization();
+        Utils.verifyUserAuthorization();
     }
 
     private void initializeFormWidgetKeyToTextMap() {
