+++ rdt-open-reader/src/main/java/io/ona/rdt/application/RDTApplication.java
@@ -1,7 +1,6 @@
 package io.ona.rdt.application;
 
 import android.app.Activity;
-import android.os.AsyncTask;
 
 import com.evernote.android.job.JobManager;
 import com.google.firebase.crashlytics.FirebaseCrashlytics;
@@ -14,7 +13,6 @@ import org.smartregister.location.helper.LocationHelper;
 import org.smartregister.receiver.SyncStatusBroadcastReceiver;
 import org.smartregister.repository.AllSharedPreferences;
 import org.smartregister.repository.Repository;
-import org.smartregister.util.SyncUtils;
 import org.smartregister.view.activity.DrishtiApplication;
 import org.smartregister.view.receiver.TimeChangedBroadcastReceiver;
 
@@ -42,9 +40,6 @@ public class RDTApplication extends DrishtiApplication {
     private Activity currentActivity;
     private RDTTestsRepository rdtTestsRepository;
     private ParasiteProfileRepository parasiteProfileRepository;
-
-    private VerifyUserAuthTask verifyUserAuthTask;
-    private SyncUtils syncUtils;
 
     public static synchronized RDTApplication getInstance() {
         return (RDTApplication) mInstance;
@@ -79,8 +74,6 @@ public class RDTApplication extends DrishtiApplication {
         if (StringUtils.isEmpty(sharedPreferences.getPreference(AllConstants.LANGUAGE_PREFERENCE_KEY))) {
             sharedPreferences.savePreference(AllConstants.LANGUAGE_PREFERENCE_KEY, BuildConfig.LOCALE);
         }
-
-        syncUtils = new SyncUtils(this);
     }
 
     private void initializeCrashlyticsAndLogging() {
@@ -159,42 +152,5 @@ public class RDTApplication extends DrishtiApplication {
             parasiteProfileRepository = new ParasiteProfileRepository();
         }
         return parasiteProfileRepository;
-    }
-
-    public void verifyUserAuthorization() {
-
-        if (verifyUserAuthTask != null) {
-            verifyUserAuthTask.cancel(true);
-        }
-
-        verifyUserAuthTask = new VerifyUserAuthTask(syncUtils);
-        verifyUserAuthTask.execute();
-    }
-
-    private static class VerifyUserAuthTask extends AsyncTask<Void, Void, Void> {
-
-        private final SyncUtils syncUtils;
-
-        VerifyUserAuthTask(SyncUtils syncUtils) {
-            this.syncUtils = syncUtils;
-        }
-
-        @Override
-        protected Void doInBackground(Void... voids) {
-
-            if (!isCancelled()) {
-
-                boolean isUserAuthorized = syncUtils.verifyAuthorization();
-                try {
-                    if (!isUserAuthorized) {
-                        syncUtils.logoutUser();
-                    }
-                } catch (Exception ex) {
-                    Timber.e(ex);
-                }
-            }
-
-            return null;
-        }
     }
 }
